Add testHelper to move out common code
