+++ src/transforms/ember-string-htmlsafe-ishtmlsafe/tests/transform.test.ts
@@ -1,37 +1,17 @@
-import {describe, test} from 'node:test'
-import assert from 'node:assert'
-import path from 'node:path'
-import fs from 'fs'
-import jscodeshift from 'jscodeshift'
+import runTestCases from '../../../utils/testHelper'
 import transform, {parser} from '#transforms/ember-string-htmlsafe-ishtmlsafe/index.js'
 
-describe('ember-string-htmlsafe-ishtmlsafe', () => {
-  const fixtureDir = path.join(__dirname, 'fixtures')
-  const testCases = [
-    {
-      name: 'should handle ember-string-htmlsafe-ishtmlsafe import and usage in javascript',
-      input: 'transform.input.js',
-      output: 'transform.output.js'
-    },
-    {
-      name: 'should handle ember-string-htmlsafe-ishtmlsafe import and usage in typescript',
-      input: 'transform.input.ts',
-      output: 'transform.output.ts'
-    }
-  ]
+const testCases = [
+  {
+    name: 'should handle ember-string-htmlsafe-ishtmlsafe import and usage in javascript',
+    input: 'transform.input.js',
+    output: 'transform.output.js'
+  },
+  {
+    name: 'should handle ember-string-htmlsafe-ishtmlsafe import and usage in typescript',
+    input: 'transform.input.ts',
+    output: 'transform.output.ts'
+  }
+]
 
-  testCases.forEach((testCase) => {
-    test(testCase.name, () => {
-      const inputPath = path.join(fixtureDir, testCase.input)
-      const outputPath = path.join(fixtureDir, testCase.output)
-      const source = fs.readFileSync(inputPath, 'utf8')
-      const expectedOutput = fs.readFileSync(outputPath, 'utf-8')
-
-      const j = jscodeshift.withParser(parser)
-
-      const actualOutput = transform({path: inputPath, source}, {j, jscodeshift: j, stats: () => {}, report: () => {}})
-
-      assert.equal(actualOutput.trim(), expectedOutput.trim())
-    })
-  })
-})
+runTestCases('ember-string-htmlsafe-ishtmlsafe', __dirname, testCases, transform, parser)
